ref: chart.tsxの責務が多すぎる問題を修正
