+++ features/population/components/chart.tsx
@@ -9,11 +9,16 @@ import {
   Legend,
 } from "chart.js";
 import { Line } from "react-chartjs-2";
-import { PopulationCategory, CHART_COLORS } from "../constants";
+import { PopulationCategory } from "../constants";
 import { SelectedPrefecture } from "../hooks";
 import { ChartDataProvider } from "./ChartDataProvider";
-import { useState, useEffect, useCallback } from "react";
-import { PopulationCompositionPerYear } from "../types";
+import { useChartData } from "../hooks/useChartData";
+import { getChartOptions } from "../config/chartConfig";
+import {
+  convertMapToList,
+  createChartDatasets,
+  extractChartLabels,
+} from "../utils/chartDataTransform";
 
 /**
  * Chartコンポーネントのプロパティ
@@ -36,14 +41,6 @@ ChartJS.register(
 );
 
 /**
- * 都道府県コードをキーとして、都道府県名と人口データを格納するマップ
- */
-type PopulationDataMap = Map<
-  number,
-  { prefName: string; data: PopulationCompositionPerYear | null }
->;
-
-/**
  * 複数の都道府県の人口推移を折れ線グラフで表示するコンポーネント
  *
  * @param props - コンポーネントのプロパティ
@@ -52,139 +49,20 @@ type PopulationDataMap = Map<
  * @returns 人口推移グラフコンポーネント
  */
 export default function Chart({ selectedPrefectures, category }: Props) {
-  const [populationDataMap, setPopulationDataMap] = useState<PopulationDataMap>(
-    new Map()
-  );
+  const { populationDataMap, handleDataLoaded } =
+    useChartData(selectedPrefectures);
 
-  // 選択された都道府県が変更されたら、マップを更新
-  useEffect(() => {
-    setPopulationDataMap((prev) => {
-      const newMap = new Map(prev);
-      const selectedCodes = new Set(selectedPrefectures.map((p) => p.prefCode));
-
-      // 選択されていない都道府県のデータを削除
-      for (const [code] of newMap) {
-        if (!selectedCodes.has(code)) {
-          newMap.delete(code);
-        }
-      }
-
-      return newMap;
-    });
-  }, [selectedPrefectures]);
-
-  /**
-   * データが読み込まれたときのコールバック関数
-   * 都道府県ごとのデータをマップに格納する
-   *
-   * @param prefCode - 都道府県コード
-   * @param prefName - 都道府県名
-   * @param data - 人口構成データ
-   */
-  const handleDataLoaded = useCallback(
-    (
-      prefCode: number,
-      prefName: string,
-      data: PopulationCompositionPerYear | null
-    ) => {
-      setPopulationDataMap((prev) => {
-        const newMap = new Map(prev);
-        newMap.set(prefCode, { prefName, data });
-        return newMap;
-      });
-    },
-    []
-  );
-
-  const options = {
-    responsive: true,
-    maintainAspectRatio: false,
-    plugins: {
-      legend: {
-        position: "top" as const,
-        labels: {
-          boxWidth: 12,
-          padding: 10,
-          font: {
-            size: 11,
-          },
-        },
-      },
-      title: {
-        display: true,
-        text: `都道府県別人口構成 - ${category}`,
-        font: {
-          size: 16,
-        },
-      },
-      tooltip: {
-        mode: "index" as const,
-        intersect: false,
-      },
-    },
-    scales: {
-      y: {
-        title: {
-          display: true,
-          text: "人口数",
-        },
-        ticks: {
-          font: {
-            size: 10,
-          },
-        },
-      },
-      x: {
-        title: {
-          display: true,
-          text: "年",
-        },
-        ticks: {
-          font: {
-            size: 10,
-          },
-          maxRotation: 45,
-          minRotation: 0,
-        },
-      },
-    },
-  };
+  // グラフのオプション設定を取得
+  const options = getChartOptions(category);
 
   // populationDataMapから配列を作成
-  const populationDataList = Array.from(populationDataMap.entries()).map(
-    ([prefCode, { prefName, data }]) => ({
-      prefCode,
-      prefName,
-      data,
-    })
-  );
+  const populationDataList = convertMapToList(populationDataMap);
 
   // データセットの作成
-  const datasets = populationDataList
-    .filter((pref) => pref.data)
-    .map((pref, index) => {
-      const categoryData = pref.data!.data.find((d) => d.label === category);
-
-      if (!categoryData) {
-        return null;
-      }
-
-      return {
-        label: pref.prefName,
-        data: categoryData.data.map((d) => d.value),
-        borderColor: CHART_COLORS[index % CHART_COLORS.length],
-        backgroundColor: CHART_COLORS[index % CHART_COLORS.length]
-          .replace("rgb", "rgba")
-          .replace(")", ", 0.5)"),
-      };
-    })
-    .filter((dataset) => dataset !== null);
+  const datasets = createChartDatasets(populationDataList, category);
 
   // X軸のラベル（年）を取得
-  const labels =
-    populationDataList.length > 0 && populationDataList[0].data
-      ? populationDataList[0].data.data[0]?.data.map((d) => d.year) || []
-      : [];
+  const labels = extractChartLabels(populationDataList);
 
   const data = {
     labels,
