+++ features/population/hooks/index.ts
@@ -1,3 +1,4 @@
 export { usePopulationData } from "./usePopulationData";
 export { usePopulationSelection } from "./usePopulationSelection";
+export { useChartData } from "./useChartData";
 export type { SelectedPrefecture } from "./usePopulationSelection";
